+++ source/extensions/omni.cae.algorithms.core/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+
+- Added support for custom glyphs.
+
 ## [1.2.0]
 
 - Changed how color map domains are set/reset. New behavior is more consistent with expectation: ranges are reset